1/nbu/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.0.13
djangorestframework==3.9.4
idna==2.8
pytz==2020.1
requests==2.21.0
urllib3==1.24.3
